+++ src/DependencyInjection/Configuration.php
@@ -6,6 +6,7 @@ use Raven_Compat;
 use Sentry\SentryBundle\EventListener\ExceptionListener;
 use Sentry\SentryBundle\EventListener\SentryExceptionListenerInterface;
 use Sentry\SentryBundle\SentrySymfonyClient;
+use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 use Symfony\Component\HttpKernel\Exception\HttpExceptionInterface;
@@ -25,6 +26,7 @@ class Configuration implements ConfigurationInterface
     public function getConfigTreeBuilder()
     {
         $treeBuilder = new TreeBuilder();
+        /** @var ArrayNodeDefinition $rootNode */
         $rootNode = $treeBuilder->root('sentry');
 
         // Basic Sentry configuration
@@ -138,10 +140,7 @@ class Configuration implements ConfigurationInterface
         return $treeBuilder;
     }
 
-    /**
-     * @return \Closure
-     */
-    private function getExceptionListenerInvalidationClosure()
+    private function getExceptionListenerInvalidationClosure(): callable
     {
         return function ($value) {
             $implements = class_implements($value);
@@ -153,10 +152,7 @@ class Configuration implements ConfigurationInterface
         };
     }
 
-    /**
-     * @return \Closure
-     */
-    private function getTrimClosure()
+    private function getTrimClosure(): callable
     {
         return function ($str) {
             $value = trim($str);